changed led control method
was using bitwise operations to update the binary int that sets which leds are on/off. Couldn't find a bitwise method to create a turn_off() that would behave as intended if called more than once.

switched to just converting the binary int to a string and then a list, then updating the list by a dictionary that maps the led name to its index in the string/list. Finally convert the list to a string and back to a binary int.